Fractions of currencies can now be bought

+++ kurios-schema.sql
@@ -42,9 +42,9 @@ CREATE TABLE transactions(
   id                SERIAL PRIMARY KEY,
   user_id           INTEGER REFERENCES users(id) on DELETE CASCADE,
   buying_id         TEXT NOT NULL,
-  buying_quantity   INTEGER NOT NULL,
+  buying_quantity   NUMERIC(18,8) NOT NULL,
   selling_id        TEXT NOT NULL, 
-  selling_quantity  INTEGER NOT NULL, 
+  selling_quantity  NUMERIC(18,8) NOT NULL, 
   created_at        TIMESTAMP NOT NULL DEFAULT NOW()
 )
